+++ src/main/java/step3/RacingGameApp.java
@@ -1,16 +1,19 @@
 package step3;
 
 import step3.model.RacingGame;
-import step3.view.View;
+import step3.view.InputView;
+import step3.view.ResultView;
 
 public class RacingGameApp {
 
     public static void main(String[] args) {
 
-        View view = new View();
+        InputView inputView = new InputView();
 
-        RacingGame racingGame = new RacingGame(view.requestStart());
+        RacingGame racingGame = new RacingGame(inputView.requestStart());
         racingGame.race();
-        view.printResult(racingGame.getResult());
+
+        ResultView resultView = new ResultView();
+        resultView.printResult(racingGame.getResult());
     }
 }
