+++ src/main/java/step3/view/InputView.java
@@ -2,20 +2,16 @@ package step3.view;
 
 import step3.model.RacingConfig;
 
-import java.util.List;
 import java.util.Scanner;
-import java.util.stream.Collectors;
 
 import static java.lang.System.out;
 
-public class View {
-
-    private static final String DISPLAY_CHARACTER = "ㅋ";
+public class InputView {
 
     private final Scanner scanner;
     private final RacingConfig racingConfig;
 
-    public View() {
+    public InputView() {
         this.scanner = new Scanner(System.in);
         int carCount = askCarCount();
         int tryCount = askTryCount();
@@ -43,13 +39,6 @@ public class View {
             throw new IllegalStateException("시작 전에 설정이 필요합니다.");
         }
         return this.racingConfig;
-    }
-
-    public void printResult(List<Integer> result) {
-        String displayResult = result.stream()
-                .map(DISPLAY_CHARACTER::repeat)
-                .collect(Collectors.joining("\n"));
-        show(displayResult);
     }
 
     private void show(String output) {
